[mms] Better generation of random IDs.
Previous method was resulting in about half the characters being
duplicated on any given system (crc32(systemname) may help produce globally
unique IDs, but it vastly reduces the number of random characters in the
string on that system).

New algorithm:
1. Take input elements and place in an array. Always add PID.
2. (Add zend_thread_id if available).
3. (Add system load averages if available.)
4. Shuffle this array to prevent identical data from being in same location in
string. (Probably not necessary given step 5 below)
5. Take MD5 sum of data and pack into binary string. (Use MD5 over SHA1
due to performance reasons - MD5 is about twice as fast, and we are not
generating cryptographically secure strings here).
6. Packed MD5 data only produces 22 character base64encoded string (since ='s
are stripped), so tack on first character of previously generated mt_rand()
to get to 23 characters.